Remove unused Components and rename Components to follow file name
