+++ webapp/src/components/course/CourseDetail/ProductCard.tsx
